fix(plugins): break uninstall policy import cycle
